+++ lib/containers/protocols/joinable.ex
@@ -1,0 +1,8 @@
+defprotocol Containers.Joinable do
+  @moduledoc """
+  Protocol for joining nested container structures by exposing the `join` function
+  """
+  @spec join(Containers.Joinable)
+  :: Containers.Joinable
+  def join(joinable)
+end